speculos/api: plug the API to the notifiers mechanism
When the API server fails to start, for example when there is already a
server bound to TCP port 5000, Speculos continued to run. This is
because the Flask web server was launched in a separate thread and
propagating a failure is not straightforward.

Introduce a new `socketpair` to signal Speculos' main event loop that
the API server stopped and that Speculos should exits.