speculos/api: always log an error when the API server raises an exception
When Flask raises an exception, the API server can close its
notification socketpair before displaying anything. If the main thread
is fast enough to quit, the API thread does not have time to display the
exception to the user, and the program appears to have closed in an
unexpected way.

In order to prevent such a synchronization issue, always report the
exception which occured in `_run`. For example if port 5000 is already
used, this leads to the following message:

    16:48:14.872:speculos.api.api: an exception occurred in the Flask
    API server: [Errno 98] Address already in use

And then the thread (which now has a proper name) displays:

    Exception in thread API-server:
    Traceback (most recent call last):
      File "/usr/lib/python3.9/threading.py", line 954, in _bootstrap_inner
        self.run()
      File "/usr/lib/python3.9/threading.py", line 892, in run
        self._target(*self._args, **self._kwargs)
    ...
      File "/usr/lib/python3.9/socketserver.py", line 466, in server_bind
        self.socket.bind(self.server_address)
    OSError: [Errno 98] Address already in use