cmake: only build tests when BUILD_TESTING is enabled
Include `CTest` in order to add `BUILD_TESTING` variable. This seems to
be a documented way to add tests in a cmake project:
https://cmake.org/cmake/help/latest/module/CTest.html#module:CTest

Put cmocka and `tests/syscalls` in `if (BUILD_TESTING)` conditional
blocks, in order to only use them when tests are being built.

When building the Python packages, C tests are not needed, so they can
be skipped, using `-DBUILD_TESTING=0`.