cmake: support linking the launcher with clang
When trying to build Speculos with clang (in order to enable more
warning flags), linking failed because clang's linker does not
understand `-Ttext-segment`:

    ld.lld: error: -Ttext-segment is not supported. Use --image-base if
    you intend to set the base address

Use `--image-base` as recommanded, with clang. Using `readelf` on the
produced `launcher` confirmed that this approach works.

While at it, use `set_property` to configure the linker options of the
launcher, instead of overridding a global variable and hoping there is
no side-effect.