setup: bump to version 0.1.0
The idea is to tag the merge commit with "0.1". Then the version of each
merge commit of the master branch can be computed with:

    LAST_TAG="$(git describe --tags --abbrev=0)"
    COUNT="$(git rev-list --count "${LAST_TAG}..HEAD")"
    VERSION="${LAST_TAG}.${COUNT}"

Introduce `tools/update_setup_version_from_git.sh` to do this in
`setup.py`. This file may then be called from GitHub Action to update
`setup.py` before creating the Python package.