main: enable using python -m speculos
`python -m speculos` runs `speculos/__main__.py` with `sys.argv`
containing `['/path/to/speculos/__main__.py']`. Add a `__main__.py` to
enable this and make the argparse parser use `speculos` as program name,
in order to have a nice name in the usage message.

This is like what Python's build module does in
https://github.com/pypa/build/blob/0.5.1/src/build/__main__.py