tools/update_setup_version_from_git: no longer rely on git tag to modify the version
As Speculos git repository does not use tags currently, the reference of
the version of the Python package is `setup.py`.

Change the way versions are computed so that it is "x.y.z" where the
`setup.py` in git contains `x.y.0` and `z` is computed by counting the
number of commits since the last change of the version line.

Suggested-in: https://github.com/LedgerHQ/speculos/issues/230