Load entire segment containing `.text`
Do this instead of just loading the section itself.

Sections are just for linking purposes and not needed at runtime with a
completely linked app. Segments (corresponding to program headers), on
the other hand, are the actual important runtime data for what should be
loaded. It is therefore more correct to load or not load whole sections
than individual segments.

We have successfully tried out various linker tricks, in order to have
things like properly initialized globals and relocations to reflect
apps install-time address on the flash. This is a small step that makes
it easier to use speculos which such experiments.

We would be happy to further improve speculos on this front, e.g. to
make it load multiple program headers provided they are adjacent, or
also load "hex" files (to be more realistic to how the "production"
workflow is without arbitrary ELF tricks). But both of these are more
involved, so we figured it was better to start with something very
simple.