speculos: Avoid flooding app with tickers and fix button and touch wait delay
This is necessary when OCR takes a lot of time:
- without this, many ticker events will be send during the OCR, which
will lead to a timeout for responding on the app side.
- as ticker events might be dropped (if speculos thread is busy), then
buttons and finger delay should take it into account.

Note that before this change, buttons and finger delay was also done
after release which is not requested and probably not wanted. This is
not the case anymore.